compiling with CFLAFS

+++ src/builtin/env.c
@@ -6,7 +6,7 @@
 /*   By: doalbaco <doalbaco@student.21-school.ru    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 22:01:54 by doalbaco          #+#    #+#             */
-/*   Updated: 2022/04/21 00:32:40 by doalbaco         ###   ########.fr       */
+/*   Updated: 2022/04/21 11:54:06 by doalbaco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static void	print_hashtable(t_hashtable *ht)
 	int		i;
 
 	i = -1;
-	while (++i < ht->size)
+	while (++i < (int) ht->size)
 	{
 		pair = ht->table[i];
 		while (pair)
